fix(studio): relocate IngestStreamEvent to @ax/lib/shared; fix latent type errors
The SPA imported axctl's ingest/stream-events.ts via a relative path that broke
on the move (type-only, so vite builds were unaffected but tsc failed). Move the
producer/consumer event contract into @ax/lib/shared as the single source;
axctl re-exports it (zero importer churn). Also fix 3 pre-existing null-safety/
cast errors the old vite-only (never-typechecked) SPA had latent.
