+++ apps/studio/src/routes/session-inspect.tsx
@@ -101,7 +101,7 @@ function totalBreakdownCost(usage: SessionTokenUsageDetail): number {
         usage.estimated_cache_creation_cost_usd,
         usage.estimated_cache_read_cost_usd,
         usage.estimated_output_cost_usd,
-    ].reduce((sum, value) => sum + (numberOrNull(value) ?? 0), 0);
+    ].reduce<number>((sum, value) => sum + (numberOrNull(value) ?? 0), 0);
 }
 
 function costBarSegments(usage: SessionTokenUsageDetail): ReadonlyArray<{ label: string; value: number | null; color: string }> {
